// src/app/(root)/admin/type.ts
export interface BlockProps {
  id: number;
  title: string;
  icon: any;
  desc: string;
  url?: string;
}

export interface UserListProps {
  id: string;
  name: string;
  regDate: string;
  img: string;
}

export interface DriverListProps {
  id: string;
  name: string;
  regDate: string;
  img: string;
}

export interface RegisterformProps {
  fullName?: string;
  userName: string;
  address: string;
  contactInfo: string;
  email: string;
  password?: string;
}

export interface AssignModalProps {
  selectOptions: any;
  assignTo?: any;
  assignDate?: Date;
  status?: string;
}